requirements.txt
pymupdf==1.19.6
openai==1.33.0
python-dotenv==1.0.1
streamlit==1.35.0
